riscv: Move IRQ code to a standalone header
arch.h is getting crowded. Move the IRQ code in a proper header file.
